+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
+
+#### [0.0.1] - 2015-09-19
+##### Added
+- Added a wrapper around Controllers
+- Added Serializer Classes for serializing request data
+- Added Renderer Classes for rendering response data
+- Added tests
+- Added tests around transpiled distribution files